move router / store to new default path
