+++ client/src/store/index.js
@@ -1,7 +1,7 @@
 import Vue from 'vue'
 import Vuex from 'vuex'
 import createPersistedState from 'vuex-persistedstate'
-import { setAuthToken } from './utils/api'
+import { setAuthToken } from '../utils/api'
 
 Vue.use(Vuex)
